fix build
use specific cabal

MUST pass buildExamples option or build will fail because of:

https://github.com/HeinrichApfelmus/reactive-banana/issues/90#issuecomment-98165162

"
The last issue looks like a problem with reactive-banana. It appears that the custom installation script tries to install an example, but fails, because the executable has not been built. I think it should work if you use the buildExamples flag:

 cabal install reactive-banana-wx -fbuildExamples

EDIT:

Actually, this is an issue with the cabal-macosx package, documented in gimbo/cabal-macosx#7 . It looks like the fix doesn't work with sandboxes, or didn't make it into the version of cabal-macosx that is currently available on hackage.
" - HeinrichApfelmus